[ADD] Added raw data message type and publisher
Updated readme 
Change node src name
