+++ README.md
@@ -18,6 +18,9 @@ The source code is released under a [MIT license](LICENSE).
 
 ## Installation
 
+### Download the latest release
+TODO
+
 ### Building from Source
 
 #### Dependencies
@@ -27,6 +30,10 @@ The source code is released under a [MIT license](LICENSE).
 - [cv_bridge]
 
 #### Building
+
+Before building from source, install [ping-protocol Python Lib](https://pypi.org/project/bluerobotics-ping/):
+
+	pip install bluerobotics-ping
 
 To build from source, clone the latest version from this repository into your catkin workspace and compile the package using
 
@@ -38,7 +45,6 @@ To build from source, clone the latest version from this repository into your ca
 ### Unit Tests
 
 TODO
-
 
 ## Usage
 
@@ -56,14 +62,20 @@ Run the main node with:
 
 ### ping360_node
 
-Creates a black and white OpenCv image using the date received from the sonar. Same as the one generated by the ping viewer.
+While continuously rotating the sonar, it publishes two types of messages:
+- The sonar's response data (the echo intensities for a given angle & range)
+- A black and white image using the date received from the sonar. Same as the one generated by the ping viewer.
 
 
 #### Published Topics
 
-* **`/ping360_sonar`** ([sensor_msgs/Image])
+* **`/sonar/images`** ([sensor_msgs/Image])
 
-	The generated sonar image.
+	The generated sonar image. 
+
+* **`/sonar/data`** ([msg/SonarEcho])
+
+	The raw sonar data.
 
 
 ## Bugs & Feature Requests
@@ -93,5 +105,6 @@ Please report bugs and request features using the [Issue Tracker](https://github
 <!-- prettier-ignore-end -->
 <!-- ALL-CONTRIBUTORS-LIST:END -->
 
-This project follows the [all-contributors](https://github.com/all-contributors/all-contributors) specification. Contributions of any kind welcome! Please refer to our [Contribution Guide](CONTRIBUTING.md)
+This project follows the [all-contributors](https://github.com/all-contributors/all-contributors) specification. 
+Contributions of any kind welcome! Please refer to our [Contribution Guide](CONTRIBUTING.md)
